requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via hikari
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via -r requirements.in
alembic==1.16.5
    # via -r requirements.in
alluka==0.3.3
    # via hikari-miru
anyio==4.10.0
    # via watchfiles
async-timeout==5.0.1
    # via
    #   aiohttp
    #   hikari-lightbulb
attrs==25.3.0
    # via
    #   aiohttp
    #   hikari
    #   hikari-miru
colorama==0.4.6
    # via
    #   colorlog
    #   hikari-miru
colorlog==6.9.0
    # via hikari
confspec==0.0.3
    # via hikari-lightbulb
exceptiongroup==1.3.0
    # via anyio
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
hikari==2.3.5
    # via
    #   -r requirements.in
    #   hikari-lightbulb
    #   hikari-miru
hikari-lightbulb==3.1.3
    # via -r requirements.in
hikari-miru==4.2.0
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   yarl
linkd==0.1.0
    # via hikari-lightbulb
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
multidict==6.6.4
    # via
    #   aiohttp
    #   hikari
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.1.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
tomli==2.2.1
    # via alembic
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   aiosqlite
    #   alembic
    #   alluka
    #   anyio
    #   exceptiongroup
    #   hikari-lightbulb
    #   multidict
    #   sqlalchemy
watchfiles==1.1.0
    # via -r requirements.in
yarl==1.20.1
    # via aiohttp
